new chatbot logic and pipeline restructuring
