+++ .gitignore
@@ -1,5 +1,4 @@
 src/data
-new_pipeline/data/
 
 # Byte-compiled / optimized / DLL files
 __pycache__/
